moving things around pushing to linux for testing next error, self still has not been fixed.
